backend/requirements.txt
Flask
requests
solana
Flask-Cors
cachetools
python-dotenv